ブランチからマージ

+++ src/Mirakurun/Server.ts
@@ -15,12 +15,11 @@
 */
 import * as fs from "fs";
 import * as http from "http";
-import { promisify } from "util";
-import express from "express";
-import cors from "cors";
-import mime from "mime";
+import * as express from "express";
+import * as cors from "cors";
+import * as mime from "mime";
 import * as openapi from "express-openapi";
-import morgan from "morgan";
+import * as morgan from "morgan";
 import * as yaml from "js-yaml";
 import { OpenAPIV2 } from "openapi-types";
 import RPCServer from "jsonrpc2-ws/lib/server";
@@ -33,23 +32,14 @@ import { createRPCServer, initRPCNotifier } from "./rpc";
 
 const pkg = require("../../package.json");
 
-export class Server {
-    /** used for test */
-    testMode = false;
+class Server {
 
     private _isRunning = false;
     private _servers = new Set<http.Server>();
     private _rpcs = new Set<RPCServer>();
 
-    get isRunning() {
-        return this._isRunning;
-    }
-
-    get servers() {
-        return this._servers;
-    }
-
     async init() {
+
         if (this._isRunning === true) {
             throw new Error("Server is running");
         }
@@ -57,37 +47,37 @@ export class Server {
 
         const serverConfig = _.config.server;
 
-        const addresses: string[] = [];
+        let addresses: string[] = [];
 
         if (serverConfig.path) {
             addresses.push(serverConfig.path);
         }
 
-        if (typeof serverConfig.port === "number") {
-            if (!this.testMode) {
-                while (true) {
-                    try {
-                        const systemIPv4s = system.getIPv4AddressesForListen();
-                        if (systemIPv4s.length > 0) {
-                            addresses.push(...systemIPv4s);
-                            break;
-                        }
-                    } catch (e) {
-                        console.error(e);
+        if (serverConfig.port) {
+            while (true) {
+                try {
+                    if (system.getIPv4AddressesForListen().length > 0) {
+                        break;
                     }
-                    log.warn("Server hasn't detected IPv4 addresses...");
-                    await sleep(5000);
+                } catch (e) {
+                    console.error(e);
                 }
+                log.warn("Server hasn't detected IPv4 addresses...");
+                await sleep(5000);
             }
 
-            addresses.push("127.0.0.1");
+            addresses = [
+                ...addresses,
+                ...system.getIPv4AddressesForListen(),
+                "127.0.0.1"
+            ];
 
             if (serverConfig.disableIPv6 !== true) {
-                if (!this.testMode) {
-                    addresses.push(...system.getIPv6AddressesForListen());
-                }
-
-                addresses.push("::1");
+                addresses = [
+                    ...addresses,
+                    ...system.getIPv6AddressesForListen(),
+                    "::1"
+                ];
             }
         }
 
@@ -96,6 +86,19 @@ export class Server {
         app.disable("x-powered-by");
         app.disable("etag");
 
+        const corsOptions: cors.CorsOptions = {
+            origin: (origin, callback) => {
+                if (!origin) {
+                    return callback(null, true);
+                }
+                if (system.isPermittedHost(origin, serverConfig.hostname)) {
+                    return callback(null, true);
+                }
+                return callback(new Error("Not allowed by CORS"));
+            }
+        };
+        app.use(cors(corsOptions));
+
         app.use(morgan(":remote-addr :remote-user :method :url HTTP/:http-version :status :res[content-length] - :response-time ms :user-agent", {
             stream: log.event as any
         }));
@@ -103,14 +106,14 @@ export class Server {
         app.use(express.json());
 
         app.use((req: express.Request, res: express.Response, next) => {
+
             if (req.ip && system.isPermittedIPAddress(req.ip) === false) {
                 req.socket.end();
                 return;
             }
 
-            const origin = req.get("Origin");
-            if (origin !== undefined) {
-                if (!system.isPermittedHost(origin, serverConfig.hostname) && !serverConfig.allowOrigins.includes(origin)) {
+            if (req.get("Origin") !== undefined) {
+                if (!system.isPermittedHost(req.get("Origin"), serverConfig.hostname)) {
                     res.status(403).end();
                     return;
                 }
@@ -123,20 +126,9 @@ export class Server {
                 }
             }
 
-            if (serverConfig.allowPNA && req.get("Access-Control-Request-Method") && req.get("Access-Control-Request-Private-Network") === "true") {
-                res.setHeader("Access-Control-Allow-Private-Network", "true");
-                res.setHeader("Private-Network-Access-Name", `Mirakurun_${serverConfig.hostname}`);
-                res.setHeader("Private-Network-Access-ID", "00:00:00:00:00:00");
-            }
-
-            res.setHeader("Cross-Origin-Resource-Policy", "cross-origin");
-
             res.setHeader("Server", "Mirakurun/" + pkg.version);
             next();
         });
-
-        // do not place before the access control
-        app.use(cors());
 
         if (!serverConfig.disableWebUI) {
             app.use(express.static("lib/ui", {
@@ -146,9 +138,8 @@ export class Server {
                     }
                 }
             }));
-            app.use("/redoc", express.static("node_modules/redoc/bundles"));
-            app.use("/redoc-try", express.static("node_modules/redoc-try/dist"));
-            app.use("/api/debug", express.static("lib/ui/redoc-ui.html"));
+            app.use("/swagger-ui", express.static("node_modules/swagger-ui-dist"));
+            app.use("/api/debug", express.static("lib/ui/swagger-ui.html"));
         }
 
         const api = yaml.load(fs.readFileSync("api.yml", "utf8")) as OpenAPIV2.Document;
@@ -162,6 +153,7 @@ export class Server {
         });
 
         app.use((err, req, res: express.Response, next) => {
+
             if (err.message === "Not allowed by CORS") {
                 res.status(403).end();
                 return;
@@ -185,71 +177,43 @@ export class Server {
             next();
         });
 
-        if (!this._isRunning) {
-            return;
-        }
+        addresses.forEach(address => {
 
-        for (const address of addresses) {
             const server = http.createServer(app);
+
             server.timeout = 1000 * 15; // 15 sec.
 
-            this._servers.add(server);
-            this._rpcs.add(createRPCServer(server));
-
-            if (regexp.unixDomainSocket.test(address)) {
-                if (fs.existsSync(address)) {
+            if (regexp.unixDomainSocket.test(address) === true || regexp.windowsNamedPipe.test(address) === true) {
+                if (process.platform !== "win32" && fs.existsSync(address) === true) {
                     fs.unlinkSync(address);
                 }
 
-                await new Promise<void>(resolve => {
-                    server.listen(address, () => {
-                        log.info("listening on http+unix://%s", address.replace(/\//g, "%2F"));
-                        resolve();
-                    });
+                server.listen(address, () => {
+                    log.info("listening on http+unix://%s", address.replace(/\//g, "%2F"));
                 });
 
-                fs.chmodSync(address, "777");
+                if (process.platform !== "win32") {
+                    fs.chmodSync(address, "777");
+                }
             } else {
-                await new Promise<void>(resolve => {
-                    server.listen(serverConfig.port, address, () => {
-                        const serverAddr = server.address();
-                        const port = typeof serverAddr === "string" ? serverConfig.port : serverAddr.port;
-                        if (address.includes(":")) {
-                            const [addr, iface] = address.split("%");
-                            log.info("listening on http://[%s]:%d (%s)", addr, port, iface);
-                        } else {
-                            log.info("listening on http://%s:%d", address, port);
-                        }
-                        resolve();
-                    });
+                server.listen(serverConfig.port, address, () => {
+                    if (address.includes(":") === true) {
+                        const [addr, iface] = address.split("%");
+                        log.info("listening on http://[%s]:%d (%s)", addr, serverConfig.port, iface);
+                    } else {
+                        log.info("listening on http://%s:%d", address, serverConfig.port);
+                    }
                 });
             }
-        }
+
+            this._servers.add(server);
+            this._rpcs.add(createRPCServer(server));
+        });
 
         // event notifications for RPC
         initRPCNotifier(this._rpcs);
 
         log.info("RPC interface is enabled");
-    }
-
-    async deinit() {
-        if (this._isRunning === false) {
-            return;
-        }
-
-        for (const rpc of this._rpcs) {
-            await rpc.close();
-        }
-
-        for (const server of this._servers) {
-            const serverCloseAsync = promisify(server.close).bind(server);
-            await serverCloseAsync();
-        }
-
-        this._rpcs.clear();
-        this._servers.clear();
-
-        this._isRunning = false;
     }
 }
